Trying to split cluster and server manage logic
