+++ handlers/admin/clusters/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ("router",)
+
+from clusters_handler import router
